+++ src/components/tabs/__internal__/tabs-header/tabs-header.component.tsx
@@ -1,15 +1,31 @@
-import React from "react";
+import React, { useRef, useState } from "react";
 import {
   StyledTabsHeaderWrapper,
   StyledTabsHeaderList,
+  StyledTabsBottomBorder,
+  StyledTabsWrapper,
+  StyledTabsBottomBorderWrapper,
 } from "./tabs-header.style";
+import useThrottle from "../../../../hooks/__internal__/useThrottle";
+
+/*  In the original prototype the tabs have shadows that fade out as you scroll horizontally.
+ *  This value is the closest replication to the way that the shadow disappears.
+ *  It is ultimately tied to the position of the scroll that will then fade the shadow in and out. */
+const fullOpacityThreshold = 128;
+
+const getOpacityRatio = (value: number) => value / fullOpacityThreshold;
+
+const getScrollRight = ({
+  scrollWidth,
+  clientWidth,
+  scrollLeft,
+}: HTMLDivElement) => scrollWidth - clientWidth - scrollLeft;
 
 export interface TabHeaderProps {
   role?: string;
   position?: "top" | "left";
   extendedLine?: boolean;
   noRightBorder?: boolean;
-  alternateStyling?: boolean;
   isInSidebar?: boolean;
   children: React.ReactNode;
   align?: "left" | "right";
@@ -21,10 +37,34 @@ const TabsHeader = ({
   position = "top",
   role,
   extendedLine,
-  alternateStyling,
   noRightBorder = false,
   isInSidebar = false,
 }: TabHeaderProps) => {
+  const [leftScrollOpacity, setLeftScrollOpacity] = useState(0);
+  const [rightScrollOpacity, setRightScrollOpacity] = useState(1);
+
+  const ref = useRef<HTMLDivElement>(null);
+
+  let isScrollable = false;
+
+  const { current } = ref;
+
+  if (position === "top" && current) {
+    isScrollable = current.scrollWidth > current.clientWidth;
+  }
+
+  const handleScroll = (
+    e: React.UIEvent<HTMLDivElement> & { target: HTMLDivElement }
+  ) => {
+    const { scrollLeft } = e.target;
+    const scrollRight = getScrollRight(e.target);
+
+    setLeftScrollOpacity(Math.min(getOpacityRatio(scrollLeft), 1));
+    setRightScrollOpacity(Math.min(getOpacityRatio(scrollRight), 1));
+  };
+
+  const throttledHandleScroll = useThrottle(handleScroll, 50);
+
   return (
     <StyledTabsHeaderWrapper isInSidebar={isInSidebar} position={position}>
       <StyledTabsHeaderList
@@ -32,11 +72,24 @@ const TabsHeader = ({
         position={position}
         role={role}
         extendedLine={extendedLine}
-        alternateStyling={alternateStyling}
         noRightBorder={noRightBorder}
         isInSidebar={isInSidebar}
+        onScroll={throttledHandleScroll}
+        leftScrollOpacity={leftScrollOpacity}
+        rightScrollOpacity={rightScrollOpacity}
+        isScrollable={isScrollable}
+        ref={ref}
       >
-        {children}
+        {position === "top" ? (
+          <StyledTabsWrapper>
+            <StyledTabsBottomBorderWrapper>
+              <StyledTabsBottomBorder />
+            </StyledTabsBottomBorderWrapper>
+            {children}
+          </StyledTabsWrapper>
+        ) : (
+          children
+        )}
       </StyledTabsHeaderList>
     </StyledTabsHeaderWrapper>
   );